% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{do_both_haps}
\alias{do_both_haps}
\title{Import both haplotypes}
\usage{
do_both_haps(root)
}
\arguments{
\item{root}{Root name of the file, without ".hapX.paf"}
}
\value{
Data frame with the realigned data
}
\description{
Import both haplotypes
}
\examples{
\dontrun{ }
}
