% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{make_chr}
\alias{make_chr}
\title{Make the output of do_both_haps into chromosomes}
\usage{
make_chr(df)
}
\arguments{
\item{df}{A data frame with with at least:
\itemize{
\item tname
\item qlen
\item tstart
\item tend
\item flip
\item qlen
}}
}
\value{
A data frame with the contigs spanning the chromosomes with minimal gaps
}
\description{
Make the output of do_both_haps into chromosomes
}
\examples{
\dontrun{ }
}
