% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{minimize_gaps}
\alias{minimize_gaps}
\title{Title}
\usage{
minimize_gaps(contigs)
}
\arguments{
\item{contigs}{Data frame with the following fields:
\itemize{
\item qname
\item start
\item end
\item str (should it be flipped?)
\item contiglen
}}
}
\value{
A data frame with the selected contigs in the order that minimizes gaps
}
\description{
Title
}
\examples{
\dontrun{ }
}
