% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read_and_flip}
\alias{read_and_flip}
\title{Read a PAF file and flip to match reference orientation}
\usage{
read_and_flip(fname)
}
\arguments{
\item{fname}{Root of file name}
}
\value{
A data frame with most of the helpful PAF columns retained
}
\description{
Read a PAF file and flip to match reference orientation
}
\examples{
\dontrun{ }
}
