% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Wrapped to run the entire pipeline}
\usage{
run_pipeline(fname)
}
\arguments{
\item{fname}{Root file name}
}
\value{
The oriented, minimal spanning contigs, taken from both haplotypes
}
\description{
Wrapped to run the entire pipeline
}
\examples{
\dontrun{ }
}
